refactor(wasm): move cljrs-dom into cljrs-wasm::dom
Consolidation stage 5, item 5. `cljrs-dom` had exactly one consumer —
`cljrs-wasm`, which called `set_globals` and `register` from `Repl::new` — no
independent artifact, and no boundary rule to stand on. Its four source files
move with `git mv` into `crates/cljrs-wasm/src/dom/`; the only edits to moved
code are the `crate::` -> `crate::dom::` path rewrites and the module doc.

Its `wasm32`-gated browser bindings (`js-sys`, `web-sys` with its feature list)
move into `cljrs-wasm` under the same target predicate, so a native build of
`cljrs-wasm` still compiles none of them and `register` is still the same
no-op there. `wasm-bindgen`, `wasm-bindgen-futures`, `getrandom` and `uuid`
were already declared by `cljrs-wasm` with the same features and are not
duplicated.

The `cljrs-dom` README's `cljrs.dom` namespace reference — the actual user
documentation for the DOM API — is folded into the `cljrs-wasm` README rather
than deleted, and `docs/replicant.md`'s source pointers are updated.

Verified with `cargo check -p cljrs-wasm --target wasm32-unknown-unknown`
(the target this module's bodies actually compile for) as well as the native
workspace build.

Package count: 30 -> 29.
