+++ Cargo.toml
@@ -28,7 +28,6 @@ members = [
     "crates/cljrs-nrepl",
     "crates/cljrs-base64",
     "crates/cljrs-dylib",
-    "crates/cljrs-dom",
     "crates/cljrs-tx",
 ]
 resolver = "2"
@@ -62,7 +61,6 @@ cljrs-lsp          = { path = "crates/cljrs-lsp",          version = "0.1.0" }
 cljrs-nrepl        = { path = "crates/cljrs-nrepl",        version = "0.1.0" }
 cljrs-base64       = { path = "crates/cljrs-base64",       version = "0.1.0" }
 cljrs-dylib        = { path = "crates/cljrs-dylib",        version = "0.1.0" }
-cljrs-dom          = { path = "crates/cljrs-dom",          version = "0.1.0" }
 cljrs-tx           = { path = "crates/cljrs-tx",           version = "0.1.0" }
 
 tracing            = "0.1.44"
