--- crates/cljrs-dom/Cargo.toml
@@ -1,50 +1,0 @@
-[package]
-name = "cljrs-dom"
-version = "0.1.0"
-edition = "2024"
-description = "Clean DOM interaction API for WASM targets (cljrs.dom namespace)"
-license.workspace = true
-repository.workspace = true
-
-[lib]
-crate-type = ["rlib"]
-
-[dependencies]
-cljrs-value  = { workspace = true }
-cljrs-gc     = { workspace = true }
-cljrs-env    = { workspace = true }
-cljrs-async  = { workspace = true }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-wasm-bindgen         = "0.2"
-wasm-bindgen-futures = "0.4"
-js-sys               = "0.3"
-# Activate the JS entropy source for transitive deps that use getrandom / uuid.
-getrandom = { version = "0.4", features = ["wasm_js"] }
-uuid      = { version = "1.22.0", features = ["v4", "js"] }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies.web-sys]
-version  = "0.3"
-features = [
-    "Window",
-    "Document",
-    "Element",
-    "HtmlElement",
-    "Node",
-    "NodeList",
-    "Text",
-    "EventTarget",
-    "Event",
-    "MouseEvent",
-    "KeyboardEvent",
-    "InputEvent",
-    "CssStyleDeclaration",
-    "HtmlInputElement",
-    "HtmlSelectElement",
-    "HtmlTextAreaElement",
-    "HtmlOptionElement",
-    "DomTokenList",
-    "HtmlHeadElement",
-    "AddEventListenerOptions",
-    "console",
-]
