+++ crates/cljrs-wasm/src/dom/mod.rs
@@ -1,8 +1,12 @@
-//! `cljrs-dom` — clean DOM interaction API for WASM targets.
+//! Clean DOM interaction API for WASM targets.
 //!
 //! Registers the `cljrs.dom` namespace with idiomatic Clojure-flavoured
 //! wrappers around the browser DOM: kebab-case names, `!`-suffixed
 //! mutations, events as Clojure maps, and `core.async` channel support.
+//!
+//! The bodies are `#[cfg(target_arch = "wasm32")]`: on a native build [`register`]
+//! is a no-op so the crate still compiles and tests, which is how the workspace
+//! builds it outside a browser target.
 #![allow(clippy::type_complexity)]
 // `CljChannel` is not `Send + Sync`, but wasm32 is single-threaded so wrapping
 // it in `Arc` (as the core.async layer does) is sound here.
